refactor(ui): rename slices directory to features and make city subdirectory
